rflib: Introduce TLV, Action, Match, Option types
TLV is a type/length/value structure that supports converting to/from
BSON. It has constructors based on a pointer or different sized integer
values.

Match,Action,Option are subtypes of TLV that restrict the values to be
stored inside the TLV to specific types and values. They also provide
functions to handle converting vectors of those types to/from BSON.

Signed-off-by: Joe Stringer <joestringernz@gmail.com>

---
v7:
- Add TLV constructors based on MACAddress,IPAddress objects
v6:
- Introduce new getters for uint16,32,64,ip_match,ip6_match
- Add "Option" TLV for storing things like priority,timeouts,ct_id
- Add TLV::optional() to determine whether the TLV can be ignored
- Implement operators
- Misc fixes