rflib: Introduce TLV, Action, Match types
TLV is a type/length/value structure that supports converting to/from
BSON. It has constructors based on a pointer or a 32-bit value.

Match,Action are subtypes of TLV that restrict the values to be stored
inside the TLV to specific types and values. They also provide functions
to handle converting vectors of Matches,Actions to/from BSON.

Signed-off-by: Joe Stringer <joestringernz@gmail.com>
---
v2:
- Add Match(TLV), Action(TLV) constructors
- Add static TLV::from_BSON() factory
- Add TLV::getLength()
- Move TLVList to ActionList,MatchList
- Make TLV::toString() use virtual type_to_string() functions
- Remove redundant constructors
- Fixed segfault on Match,Action construction
- Do basic sanity checking on BSON input
- Set output action length correctly
- Serialise 32-bit values correctly
- Serialised TLVs now always have "value" (can be 0-length)
- Comment the TLV class
- Rename "data" to "value"

Signed-off-by: Joe Stringer <joestringernz@gmail.com>