Add support for FPM interface
Forwarding Path Manager (FPM) interface is a new protocol for allowing
routing engines to send RIB updates in a platform independent way. This
is an alternative to NetLink.

This patch adds a new thread to liaise with software that supports
having an external Forwarding Path Manager (FPM). To compile with FPM
support, see the Makefile in rfclient. FPM is disabled by default.