Adds fastpath, currently this will work with normal ports and inter-switch links
The idea of fastpath is to create a direct link to the controller in the data-plane
and send packets to the controller this way. This avoids the overhead of encapsulation
on PKTIN and PKTOUT packets. Fastpath can be performed entirely on the switching
hardware making it less subceptable to dos attacks.

To use this you'll need to create a rffastpath.csv with the following columns
ct_id,dp_id,dp_port,dp0_port
And the corresponding rfdp0links.csv with the following format
device,dp0_port

Here the device is the physical interface and the dp0_port is used to verify against.
Note the port number it'll get is based on the next avaliable on dp0. As such
this should be set to the number of ports already on dp0 and count up. This should
match with dp0_port specified in rffastpath.csv

Once configured shortest path will be used to find a fastpath to the controller from
every other dataplane element. There can be multiple fastpath links, however no
attempt is made to load balance beyond shortest path.

Limitations:
This currently requires multitable support.
Fastpaths use VLANs and as such the network may not be vlan transparent. If VLAN is needed pcp could be set to identify fastpath packets etc.
Only the default port config file and isl is parsed and this likely wont work with other links.